Registry is now a Mixin, not a Class
RedisRegistry no longer inherits methods from the memory-based Registry
class. Methods with common implementation are stored in the Registry
module, and the contents of the class have moved to MemoryRegistry
